@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.5+7
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.22.6"
 
 dependencies:
@@ -26,31 +26,30 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_cupertino_localizations: ^1.0.1
   intl: ^0.17.0
-  devicelocale: ^0.5.0
-  shared_preferences: ^2.0.8
-  provider: ^6.0.1
-  path_provider: ^2.0.5
-  mobx: ^2.0.5
-  flutter_mobx: 2.0.2
-  hive: ^2.0.4
+  devicelocale: ^0.6.0
+  shared_preferences: ^2.1.0
+  provider: ^6.0.5
+  path_provider: ^2.0.14
+  mobx: ^2.1.4
+  flutter_mobx: ^2.0.6+5
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  http: ^0.13.4
-  yaml: ^3.1.0
+  http: ^0.13.5
+  yaml: ^3.1.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.3
+  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.1.4
-  build_resolvers: ^2.0.4
-  mobx_codegen: ^2.0.4
-  hive_generator: ^1.1.1
-  flutter_launcher_icons: ^0.9.2
+  build_runner: ^2.3.3
+  build_resolvers: ^2.2.0
+  mobx_codegen: ^2.2.0
+  hive_generator: ^2.0.0
+  flutter_launcher_icons: ^0.13.0
 
 
 flutter_icons:
