@@ -9,7 +9,7 @@ import 'oxen/oxen_app_bar.dart';
 enum AppBarStyle { regular, withShadow }
 
 abstract class BasePage extends StatelessWidget {
-  String get title => null;
+  String? get title => null;
 
   bool get isModalBackButton => false;
 
@@ -23,8 +23,8 @@ abstract class BasePage extends StatelessWidget {
 
   final GlobalKey<ScaffoldState> scaffoldKey = new GlobalKey<ScaffoldState>();
 
-  Widget leading(BuildContext context) {
-    if (ModalRoute.of(context).isFirst) {
+  Widget? leading(BuildContext context) {
+    if (ModalRoute.of(context)?.isFirst == true) {
       return null;
     }
 
@@ -46,21 +46,21 @@ abstract class BasePage extends StatelessWidget {
     );
   }
 
-  Widget middle(BuildContext context) {
+  Widget? middle(BuildContext context) {
     return title == null
         ? null
         : Text(
-            title,
+            title!,
             style: TextStyle(
                 fontSize: 16.0,
                 fontWeight: FontWeight.w600,
-                color: Theme.of(context).primaryTextTheme.headline6.color),
+                color: Theme.of(context).primaryTextTheme.titleLarge?.color),
           );
   }
 
-  Widget trailing(BuildContext context) => null;
+  Widget? trailing(BuildContext context) => null;
 
-  Widget floatingActionButton(BuildContext context) => null;
+  Widget? floatingActionButton(BuildContext context) => null;
 
   ObstructingPreferredSizeWidget appBar(BuildContext context) {
     final _themeChanger = Provider.of<ThemeChanger>(context);
@@ -74,7 +74,7 @@ abstract class BasePage extends StatelessWidget {
             middle: middle(context),
             trailing: trailing(context),
             backgroundColor: _isDarkTheme
-                ? Theme.of(context).backgroundColor
+                ? Theme.of(context).colorScheme.background
                 : backgroundColor);
 
       case AppBarStyle.withShadow:
@@ -84,7 +84,7 @@ abstract class BasePage extends StatelessWidget {
             middle: middle(context),
             trailing: trailing(context),
             backgroundColor: _isDarkTheme
-                ? Theme.of(context).backgroundColor
+                ? Theme.of(context).colorScheme.background
                 : backgroundColor);
 
       default:
@@ -94,14 +94,14 @@ abstract class BasePage extends StatelessWidget {
             middle: middle(context),
             trailing: trailing(context),
             backgroundColor: _isDarkTheme
-                ? Theme.of(context).backgroundColor
+                ? Theme.of(context).colorScheme.background
                 : backgroundColor);
     }
   }
 
   Widget body(BuildContext context);
 
-  Widget bottomNavigationBar(BuildContext context) => null;
+  Widget? bottomNavigationBar(BuildContext context) => null;
 
   @override
   Widget build(BuildContext context) {
@@ -111,7 +111,7 @@ abstract class BasePage extends StatelessWidget {
     return Scaffold(
         key: scaffoldKey,
         backgroundColor:
-            _isDarkTheme ? Theme.of(context).backgroundColor : backgroundColor,
+            _isDarkTheme ? Theme.of(context).colorScheme.background : backgroundColor,
         resizeToAvoidBottomInset: resizeToAvoidBottomPadding,
         appBar: appBar(context),
         body: SafeArea(child: body(context)),
