@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:oxen_service_node/src/widgets/nav/nav_list_trailing.dart';
 
 class NavListArrow extends StatelessWidget {
-  NavListArrow({this.text, this.leading, this.onTap});
+  NavListArrow({required this.text, this.leading, this.onTap});
 
   final String text;
-  final Widget leading;
-  final GestureTapCallback onTap;
+  final Widget? leading;
+  final GestureTapCallback? onTap;
 
   @override
   Widget build(BuildContext context) {
@@ -14,7 +14,7 @@ class NavListArrow extends StatelessWidget {
         leading: leading,
         text: text,
         trailing: Icon(Icons.arrow_forward_ios_sharp,
-            color: Theme.of(context).primaryTextTheme.headline6.color,
+            color: Theme.of(context).primaryTextTheme.titleLarge?.color,
             size: 20),
         onTap: onTap);
   }
