@@ -4,139 +4,45 @@ import 'package:oxen_service_node/src/utils/theme/palette.dart';
 
 class PrimaryButton extends StatelessWidget {
   const PrimaryButton(
-      {@required this.onPressed,
-      @required this.text,
-      @required this.color,
-      @required this.borderColor,
+      {required this.onPressed,
+      required this.text,
+      required this.color,
+      required this.borderColor,
       this.textColor,
       this.isDisabled = false,
       this.onDisabledPressed});
 
   final VoidCallback onPressed;
-  final VoidCallback onDisabledPressed;
-  final Color color;
-  final Color textColor;
-  final Color borderColor;
+  final VoidCallback? onDisabledPressed;
+  final Color? color;
+  final Color? textColor;
+  final Color? borderColor;
   final String text;
   final bool isDisabled;
 
   @override
   Widget build(BuildContext context) {
     final textColor =
-        this.textColor ?? Theme.of(context).primaryTextTheme.button.color;
+        this.textColor ?? Theme.of(context).primaryTextTheme.labelLarge?.color;
+
+    final borderColor = this.borderColor ?? Theme.of(context).primaryTextTheme.labelLarge?.decorationColor;
 
     return ButtonTheme(
         minWidth: double.infinity,
         height: 56.0,
-        child: FlatButton(
+        child: TextButton(
           onPressed: isDisabled ? onDisabledPressed : onPressed,
-          color: isDisabled ? Colors.transparent : color,
-          shape: RoundedRectangleBorder(
-              side: BorderSide(color: borderColor),
-              borderRadius: BorderRadius.circular(10.0)),
+          style: TextButton.styleFrom(
+            backgroundColor: isDisabled ? Colors.transparent : color,
+            shape: RoundedRectangleBorder(
+                side: BorderSide(color: borderColor!),
+                borderRadius: BorderRadius.circular(10.0)),
+          ),
+
           child: Text(text,
               style: TextStyle(
                   fontSize: 16.0,
                   color: isDisabled ? Palette.darkGrey : textColor)),
-        ));
-  }
-}
-
-class LoadingPrimaryButton extends StatelessWidget {
-  const LoadingPrimaryButton(
-      {@required this.onPressed,
-      @required this.text,
-      @required this.color,
-      @required this.borderColor,
-      this.isDisabled = false,
-      this.isLoading = false});
-
-  final VoidCallback onPressed;
-  final Color color;
-  final Color borderColor;
-  final bool isLoading;
-  final bool isDisabled;
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    return ButtonTheme(
-        minWidth: double.infinity,
-        height: 56.0,
-        child: FlatButton(
-          onPressed: (isLoading || isDisabled) ? null : onPressed,
-          color: color,
-          shape: RoundedRectangleBorder(
-              side: BorderSide(color: borderColor),
-              borderRadius: BorderRadius.circular(10.0)),
-          child: isLoading
-              ? CupertinoActivityIndicator(animating: true)
-              : Text(text,
-                  style: TextStyle(
-                      fontSize: 16.0,
-                      color: Theme.of(context).primaryTextTheme.button.color)),
-        ));
-  }
-}
-
-class PrimaryIconButton extends StatelessWidget {
-  const PrimaryIconButton({
-    @required this.onPressed,
-    @required this.iconData,
-    @required this.text,
-    @required this.color,
-    this.textColor,
-    @required this.borderColor,
-    @required this.iconColor,
-    @required this.iconBackgroundColor,
-  });
-
-  final VoidCallback onPressed;
-  final IconData iconData;
-  final Color color;
-  final Color textColor;
-  final Color borderColor;
-  final Color iconColor;
-  final Color iconBackgroundColor;
-  final String text;
-
-  @override
-  Widget build(BuildContext context) {
-    final textColor =
-        this.textColor ?? Theme.of(context).primaryTextTheme.button.color;
-
-    return ButtonTheme(
-        minWidth: double.infinity,
-        height: 56.0,
-        child: FlatButton(
-          onPressed: onPressed,
-          color: color,
-          shape: RoundedRectangleBorder(
-              side: BorderSide(color: borderColor),
-              borderRadius: BorderRadius.circular(10.0)),
-          child: Stack(
-            children: <Widget>[
-              Row(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: <Widget>[
-                  Container(
-                    width: 28.0,
-                    height: 56.0,
-                    decoration: BoxDecoration(
-                        shape: BoxShape.circle, color: iconBackgroundColor),
-                    child: Icon(iconData, color: iconColor, size: 22.0),
-                  ),
-                ],
-              ),
-              Container(
-                height: 56.0,
-                child: Center(
-                  child: Text(text,
-                      style: TextStyle(fontSize: 16.0, color: textColor)),
-                ),
-              )
-            ],
-          ),
         ));
   }
 }
