@@ -8,10 +8,10 @@ class ScrollableWithBottomSection extends StatefulWidget {
       this.contentPadding,
       this.bottomSectionPadding});
 
-  final Widget content;
-  final Widget bottomSection;
-  final EdgeInsets contentPadding;
-  final EdgeInsets bottomSectionPadding;
+  final Widget? content;
+  final Widget? bottomSection;
+  final EdgeInsets? contentPadding;
+  final EdgeInsets? bottomSectionPadding;
 
   @override
   ScrollableWithBottomSectionState createState() =>
