@@ -5,8 +5,8 @@ class Spinner extends StatefulWidget {
   final Duration duration;
 
   const Spinner({
-    Key key,
-    @required this.icon,
+    Key? key,
+    required this.icon,
     this.duration = const Duration(milliseconds: 1800),
   }) : super(key: key);
 
@@ -15,8 +15,8 @@ class Spinner extends StatefulWidget {
 }
 
 class _SpinnerState extends State<Spinner> with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Widget _child;
+  late AnimationController _controller;
+  late Widget _child;
 
   @override
   void initState() {
@@ -38,7 +38,7 @@ class _SpinnerState extends State<Spinner> with SingleTickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     return RotationTransition(
-      turns: ReverseAnimation(_controller),
+      turns: ReverseAnimation(_controller.view),
       child: _child,
     );
   }
